added integration test to gitlab runner

+++ par/start_Tests.sh
@@ -11,17 +11,17 @@ INTEGRATIONTEST_EXE="./../build/bin/Test_CompareMisfit"
 
 
 # Run unit tests
-#${UNITTEST_EXE}
-#if [ "$?" != "0" ]; then
-#	echo "Test failed ! "
-#	exit 1
-#fi
+${UNITTEST_EXE}
+if [ "$?" != "0" ]; then
+	echo "Test failed ! "
+	exit 1
+fi
 
  Run simulation tests
 rm -rf ci/*.ci.*
 
 mpirun -np ${NUM_MPI_PROCESSES} ${IFOS_EXE} ./ci/configuration_ci.2D.acoustic.txt
- ./../bin/Tests/Test_CompareMisfit ci/configuration_ci.2D.acoustic.txt
+ ./../build/bin/Test_CompareMisfit ci/configuration_ci.2D.acoustic.txt
 if [ "$?" != "0" ]; then
 	echo "Test failed ! "
 	return
